Initial Test and Plan clean up
