--- packages/loot-core/src/shared/derived-bounds-test.test.ts
@@ -1,115 +1,0 @@
-/**
- * TDD Test: Test derivedBounds Logic in Isolation
- *
- * This test verifies why the derivedBounds fix isn't working
- */
-
-import * as monthUtils from './months';
-
-describe('derivedBounds Logic Testing', () => {
-  beforeEach(() => {
-    monthUtils.setPayPeriodConfig({
-      enabled: true,
-      payFrequency: 'biweekly',
-      startDate: '2025-01-01',
-    });
-  });
-
-  test('ISOLATION: Test derivedBounds conversion logic', () => {
-    // This is the exact logic from index.tsx derivedBounds
-    const bounds = {
-      start: '2025-01',
-      end: '2026-09',
-    };
-
-    const convertMonthToPayPeriod = (monthStr: string) => {
-      // If already a pay period, keep it
-      const mm = parseInt(monthStr.slice(5, 7));
-      if (Number.isFinite(mm) && mm >= 13) return monthStr;
-
-      // For calendar months, convert to pay period
-      const year = parseInt(monthStr.slice(0, 4));
-      const month = parseInt(monthStr.slice(5, 7));
-
-      if (month <= 6) {
-        return `${year}-13`; // First pay period of year
-      } else {
-        return `${year}-38`; // Last pay period of year
-      }
-    };
-
-    const derivedBounds = {
-      start: convertMonthToPayPeriod(bounds.start),
-      end: convertMonthToPayPeriod(bounds.end),
-    };
-
-    expect(derivedBounds).toEqual({
-      start: '2025-13',
-      end: '2026-38',
-    });
-
-    // Test that this prevents the mixed range issue
-    const startMonth = '2025-31';
-    const endMonth = monthUtils.addMonths(startMonth, 11);
-
-    const finalStart =
-      startMonth < derivedBounds.start ? derivedBounds.start : startMonth;
-    const finalEnd =
-      endMonth > derivedBounds.end ? derivedBounds.end : endMonth;
-
-    expect(monthUtils.isPayPeriod(finalStart)).toBe(true);
-    expect(monthUtils.isPayPeriod(finalEnd)).toBe(true);
-
-    expect(() => {
-      monthUtils.rangeInclusive(finalStart, finalEnd);
-    }).not.toThrow();
-  });
-
-  test('DIAGNOSIS: Check if derivedBounds is being passed correctly', () => {
-    // Test what happens if derivedBounds isn't being used
-    const originalBounds = {
-      start: '2025-01',
-      end: '2026-09',
-    };
-
-    const derivedBounds = {
-      start: '2025-13',
-      end: '2026-38',
-    };
-
-    // Test with original bounds (should fail)
-    const startMonth = '2025-31';
-    const endMonth = monthUtils.addMonths(startMonth, 11);
-
-    // With original bounds - FAILS
-    const badFinalEnd =
-      endMonth > originalBounds.end ? originalBounds.end : endMonth;
-    expect(badFinalEnd).toBe('2026-09'); // Calendar month!
-    expect(monthUtils.isPayPeriod(badFinalEnd)).toBe(false);
-
-    // With derived bounds - WORKS
-    const goodFinalEnd =
-      endMonth > derivedBounds.end ? derivedBounds.end : endMonth;
-    expect(goodFinalEnd).toBe('2026-16'); // Pay period!
-    expect(monthUtils.isPayPeriod(goodFinalEnd)).toBe(true);
-  });
-
-  test('HYPOTHESIS: derivedBounds might not be applied in some code path', () => {
-    // Test if there's a condition where derivedBounds isn't used
-    const payPeriodViewEnabled = true;
-    const derivedStartMonth = '2025-31';
-
-    // Simulate the conditions from index.tsx
-    const config = monthUtils.getPayPeriodConfig();
-    const usePayPeriods = config?.enabled;
-
-    expect(usePayPeriods).toBe(true);
-    expect(monthUtils.isPayPeriod(derivedStartMonth)).toBe(true);
-
-    // If these conditions are met, derivedBounds should be used
-    // If derivedBounds isn't being used, the issue might be:
-    // 1. derivedBounds isn't being passed to MonthsProvider
-    // 2. There's another code path that bypasses derivedBounds
-    // 3. MonthsProvider is called before derivedBounds is calculated
-  });
-});
