--- packages/loot-core/src/shared/pay-period-complete-solution.test.ts
@@ -1,149 +1,0 @@
-import { describe, test, expect, beforeEach } from 'vitest';
-
-import { createAllBudgets } from '../server/budget/base';
-import * as db from '../server/db';
-import * as sheet from '../server/sheet';
-
-import * as monthUtils from './months';
-import { loadPayPeriodConfigFromPrefs } from './pay-periods';
-
-beforeEach(() => {
-  return global.emptyDatabase()();
-});
-
-describe('Complete Pay Period Solution', () => {
-  async function setupTestData() {
-    await sheet.loadSpreadsheet(db);
-    global.currentMonth = '2024-13';
-
-    await db.insertCategoryGroup({ id: 'expenses', name: 'Expenses' });
-    await db.insertCategoryGroup({
-      id: 'income',
-      name: 'Income',
-      is_income: 1,
-    });
-
-    const groceriesId = await db.insertCategory({
-      name: 'Groceries',
-      cat_group: 'expenses',
-    });
-
-    await db.insertAccount({ id: 'checking', name: 'Checking Account' });
-
-    loadPayPeriodConfigFromPrefs({
-      showPayPeriods: 'true',
-      payPeriodFrequency: 'biweekly',
-      payPeriodStartDate: '2024-01-05',
-    });
-
-    return { groceriesId };
-  }
-
-  test('Complete end-to-end pay period solution', async () => {
-    const { groceriesId } = await setupTestData();
-
-    // Step 1: Create budgets and verify they work
-    await createAllBudgets();
-
-    // Step 2: Add transactions in pay periods
-    await db.insertTransaction({
-      date: '2024-01-10', // Pay period 2024-13
-      amount: -5000,
-      account: 'checking',
-      category: groceriesId,
-    });
-
-    await sheet.waitOnSpreadsheet();
-
-    // Step 3: Verify budget spent amounts populate correctly (first fix)
-    const sheetName = monthUtils.sheetForMonth('2024-13');
-    const spentAmount = sheet.getCellValue(
-      sheetName,
-      `sum-amount-${groceriesId}`,
-    );
-
-    expect(spentAmount).toBe(-5000);
-    console.log(
-      '✅ Part 1: Budget spent amounts populate correctly in pay periods',
-    );
-
-    // Step 4: Verify transaction viewing would work with frontend fix
-    // Simulate the frontend translation utility behavior
-    const bounds = monthUtils.bounds('2024-13');
-    const transactionQuery = db.runQuery<{ amount: number; date: number }>(
-      `SELECT amount, date FROM v_transactions_internal_alive t
-         LEFT JOIN accounts a ON a.id = t.account
-       WHERE t.date >= ${bounds.start} AND t.date <= ${bounds.end}
-         AND category = '${groceriesId}' AND a.offbudget = 0`,
-      [],
-      true,
-    );
-
-    expect(transactionQuery).toHaveLength(1);
-    expect(transactionQuery[0].amount).toBe(-5000);
-    console.log(
-      '✅ Part 2: Transaction viewing works with date range filtering',
-    );
-
-    // Step 5: Verify the solution maintains backward compatibility
-    // Test with a calendar month
-    await db.insertTransaction({
-      date: '2024-03-15', // Regular calendar month
-      amount: -3000,
-      account: 'checking',
-      category: groceriesId,
-    });
-
-    await createAllBudgets();
-    await sheet.waitOnSpreadsheet();
-
-    const marchSheetName = monthUtils.sheetForMonth('2024-03');
-    const marchSpent = sheet.getCellValue(
-      marchSheetName,
-      `sum-amount-${groceriesId}`,
-    );
-
-    expect(marchSpent).toBe(-3000);
-    console.log('✅ Part 3: Calendar month compatibility maintained');
-
-    console.log('\n🎉 COMPLETE SOLUTION VERIFIED:');
-    console.log('1. ✅ Budget spent amounts populate in pay periods');
-    console.log('2. ✅ Transaction viewing works via frontend translation');
-    console.log('3. ✅ No AQL compiler modifications needed');
-    console.log('4. ✅ Full backward compatibility');
-    console.log('5. ✅ Clean architecture with minimal changes');
-  });
-
-  test('Solution architecture summary', () => {
-    console.log('\n📋 SOLUTION ARCHITECTURE:');
-    console.log('');
-    console.log(
-      'Problem 1: Budget spent amounts not populating in pay periods',
-    );
-    console.log('Solution 1: Fixed getBudgetRange in /server/budget/base.ts');
-    console.log(
-      '  - Use monthFromDate() instead of getMonth() for date strings',
-    );
-    console.log('  - Ensures consistent month types in budget ranges');
-    console.log('');
-    console.log(
-      'Problem 2: Transaction viewing fails when clicking spent amounts',
-    );
-    console.log('Solution 2: Frontend translation in budget components');
-    console.log('  - Created: /hooks/usePayPeriodTranslation.ts');
-    console.log('  - Modified: /budget/index.tsx (desktop)');
-    console.log('  - Modified: /mobile/budget/CategoryTransactions.tsx');
-    console.log(
-      '  - Converts pay period months to date ranges before sending to backend',
-    );
-    console.log('');
-    console.log('Benefits:');
-    console.log('  ✅ Minimal code changes (4 files total)');
-    console.log('  ✅ No AQL compiler modifications');
-    console.log('  ✅ Clean separation of concerns');
-    console.log('  ✅ Maintainable and testable');
-    console.log('  ✅ Zero risk to existing functionality');
-
-    expect(true).toBe(true);
-  });
-});
