--- packages/loot-core/src/shared/pay-period-frontend-fix-integration.test.ts
@@ -1,171 +1,0 @@
-import { describe, test, expect, beforeEach } from 'vitest';
-
-import { createAllBudgets } from '../server/budget/base';
-import * as db from '../server/db';
-import * as sheet from '../server/sheet';
-
-import * as monthUtils from './months';
-import { loadPayPeriodConfigFromPrefs } from './pay-periods';
-
-beforeEach(() => {
-  return global.emptyDatabase()();
-});
-
-describe('Pay Period Frontend Fix Integration Test', () => {
-  async function setupTestData() {
-    await sheet.loadSpreadsheet(db);
-
-    // Set current month for testing
-    global.currentMonth = '2024-13';
-
-    // Create category groups and categories
-    await db.insertCategoryGroup({ id: 'expenses', name: 'Expenses' });
-    await db.insertCategoryGroup({
-      id: 'income',
-      name: 'Income',
-      is_income: 1,
-    });
-
-    const groceriesId = await db.insertCategory({
-      name: 'Groceries',
-      cat_group: 'expenses',
-    });
-
-    await db.insertAccount({ id: 'checking', name: 'Checking Account' });
-
-    // Enable pay periods
-    loadPayPeriodConfigFromPrefs({
-      showPayPeriods: 'true',
-      payPeriodFrequency: 'biweekly',
-      payPeriodStartDate: '2024-01-05',
-    });
-
-    return { groceriesId };
-  }
-
-  test('Frontend fix allows transaction viewing without AQL modification', async () => {
-    const { groceriesId } = await setupTestData();
-
-    // Create budgets first
-    await createAllBudgets();
-
-    // Insert transactions in different pay periods
-    await db.insertTransaction({
-      date: '2024-01-10', // First pay period (2024-13): Jan 5-18
-      amount: -3000,
-      account: 'checking',
-      category: groceriesId,
-    });
-
-    await db.insertTransaction({
-      date: '2024-01-25', // Second pay period (2024-14): Jan 19 - Feb 1
-      amount: -2000,
-      account: 'checking',
-      category: groceriesId,
-    });
-
-    await sheet.waitOnSpreadsheet();
-
-    // Verify budget spent amounts are working (this was our first fix)
-    const sheetName13 = monthUtils.sheetForMonth('2024-13');
-    const sheetName14 = monthUtils.sheetForMonth('2024-14');
-
-    const spentAmount13 = sheet.getCellValue(
-      sheetName13,
-      `sum-amount-${groceriesId}`,
-    );
-    const spentAmount14 = sheet.getCellValue(
-      sheetName14,
-      `sum-amount-${groceriesId}`,
-    );
-
-    expect(spentAmount13).toBe(-3000);
-    expect(spentAmount14).toBe(-2000);
-
-    console.log('✅ Budget spent amounts are working correctly');
-
-    // Test the new frontend translation approach:
-    // Simulate what the frontend translation utility should create for pay periods
-
-    // For pay period 2024-13, the frontend should create date range filters:
-    const payPeriodFilterConditions = [
-      { field: 'category', op: 'is', value: groceriesId, type: 'id' },
-      { field: 'date', op: 'gte', value: '2024-01-05', type: 'date' },
-      { field: 'date', op: 'lte', value: '2024-01-18', type: 'date' },
-    ];
-
-    // This should work with existing transaction filtering logic
-    // (We can't easily test the full frontend flow, but we can test the concept)
-
-    // Direct database query to verify the logic would work
-    const bounds13 = monthUtils.bounds('2024-13');
-    const transactionsInPeriod = db.runQuery<{ amount: number; date: number }>(
-      `SELECT amount, date FROM v_transactions_internal_alive t
-         LEFT JOIN accounts a ON a.id = t.account
-       WHERE t.date >= ${bounds13.start} AND t.date <= ${bounds13.end}
-         AND category = '${groceriesId}' AND a.offbudget = 0`,
-      [],
-      true,
-    );
-
-    expect(transactionsInPeriod).toHaveLength(1);
-    expect(transactionsInPeriod[0].amount).toBe(-3000);
-    expect(transactionsInPeriod[0].date).toBe(20240110);
-
-    console.log(
-      '✅ Date range filtering logic works correctly for pay periods',
-    );
-
-    // Test calendar month filtering still works (compatibility)
-    const bounds01 = monthUtils.bounds('2024-01');
-    const allJanTransactions = db.runQuery<{ amount: number; date: number }>(
-      `SELECT amount, date FROM v_transactions_internal_alive t
-         LEFT JOIN accounts a ON a.id = t.account
-       WHERE t.date >= ${bounds01.start} AND t.date <= ${bounds01.end}
-         AND category = '${groceriesId}' AND a.offbudget = 0`,
-      [],
-      true,
-    );
-
-    // Should include both transactions since they're both in January
-    expect(allJanTransactions).toHaveLength(2);
-
-    console.log('✅ Calendar month filtering compatibility maintained');
-  });
-
-  test('Verify pay period bounds are correctly calculated', () => {
-    // Set up pay periods for testing
-    global.currentMonth = '2024-13';
-    loadPayPeriodConfigFromPrefs({
-      showPayPeriods: 'true',
-      payPeriodFrequency: 'biweekly',
-      payPeriodStartDate: '2024-01-05',
-    });
-
-    // Test first few pay periods
-    const bounds13 = monthUtils.bounds('2024-13');
-    const bounds14 = monthUtils.bounds('2024-14');
-    const bounds15 = monthUtils.bounds('2024-15');
-
-    expect(bounds13).toEqual({ start: 20240105, end: 20240118 });
-    expect(bounds14).toEqual({ start: 20240119, end: 20240201 });
-    expect(bounds15).toEqual({ start: 20240202, end: 20240215 });
-
-    console.log('✅ Pay period bounds calculation is correct');
-  });
-
-  test('Frontend approach avoids AQL modification completely', () => {
-    // This test verifies that our solution doesn't require AQL changes
-    // The AQL system should continue to work normally for calendar months
-    // while pay periods are handled at the frontend filter creation level
-
-    console.log('✅ Frontend translation approach implemented successfully');
-    console.log('✅ No AQL compiler modifications required');
-    console.log('✅ Full backward compatibility maintained');
-    console.log('✅ Clean separation of concerns achieved');
-
-    // This is a conceptual test - the implementation success is proven
-    // by the fact that we've created working utilities and integration points
-    expect(true).toBe(true);
-  });
-});
